+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes will be documented in this file
 ## [0.0.2](https://github.com/pablobascunana/youml-manager/compare/d51b3d7...develop) (WIP)
 
 ### Added
-* [PR-6](https://github.com/pablobascunana/youml-manager/pull/6) Add RabbitMQ docker compose
-* [PR-7](https://github.com/pablobascunana/youml-manager/pull/7) Create Django project
+* [PR-24](https://github.com/pablobascunana/youml-manager/pull/19) Create RabbitMQ consumer command
 * [PR-9](https://github.com/pablobascunana/youml-manager/pull/9) Create JWT token security and add decorator to secure views
+* [PR-7](https://github.com/pablobascunana/youml-manager/pull/7) Create Django project
+* [PR-6](https://github.com/pablobascunana/youml-manager/pull/6) Add RabbitMQ docker compose
 
 ## [0.0.1](https://github.com/pablobascunana/youml-manager/compare/33cfc97...d51b3d7) (23-11-2022)